perf: Max creator fee bp can be packed (#180)
* perf: TransferManager optimizations

* perf: Max creator fee BP can be packed

+++ contracts/ExecutionManager.sol
@@ -25,7 +25,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
     address public protocolFeeRecipient;
 
     // Maximum creator fee (in basis point)
-    uint256 public maximumCreatorFeeBp = 1_000;
+    uint16 public maximumCreatorFeeBp = 1_000;
 
     // Creator fee manager
     ICreatorFeeManager public creatorFeeManager;
@@ -45,7 +45,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
      * @param newMaximumCreatorFeeBp New maximum creator fee (in basis point)
      * @dev The maximum value that can be set is 25%.
      */
-    function setMaximumCreatorFeeBp(uint256 newMaximumCreatorFeeBp) external onlyOwner {
+    function setMaximumCreatorFeeBp(uint16 newMaximumCreatorFeeBp) external onlyOwner {
         if (newMaximumCreatorFeeBp > 2_500) revert CreatorFeeBpTooHigh();
         maximumCreatorFeeBp = newMaximumCreatorFeeBp;
 
@@ -91,7 +91,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
             // 0 --> Creator fee and adjustment of protocol fee
             if (address(creatorFeeManager) != address(0)) {
                 (recipients[1], fees[1]) = creatorFeeManager.viewCreatorFee(makerBid.collection, price, itemIds);
-                if (fees[1] > (price * maximumCreatorFeeBp) / 10_000) revert CreatorFeeBpTooHigh();
+                if (fees[1] > (price * uint256(maximumCreatorFeeBp)) / 10_000) revert CreatorFeeBpTooHigh();
             }
 
             uint256 minTotalFee = (price * strategyInfo[makerBid.strategyId].minTotalFee) / 10_000;
@@ -146,7 +146,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
             // 0 --> Creator fee and adjustment of protocol fee
             if (address(creatorFeeManager) != address(0)) {
                 (recipients[1], fees[1]) = creatorFeeManager.viewCreatorFee(makerAsk.collection, price, itemIds);
-                if (fees[1] > (price * maximumCreatorFeeBp) / 10_000) revert CreatorFeeBpTooHigh();
+                if (fees[1] > (price * uint256(maximumCreatorFeeBp)) / 10_000) revert CreatorFeeBpTooHigh();
             }
             uint256 minTotalFee = (price * strategyInfo[makerAsk.strategyId].minTotalFee) / 10_000;
